simplified_auth_anomaly_detection/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
flask>=2.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
pyyaml>=6.0
datetime